feat: implement admin-only product management functions with authentication checks and refactor product actions
