+++ lib/actions/product-actions.ts
@@ -1,11 +1,8 @@
 'use server';
 
 import { prisma } from '@/db/prisma';
-import { convertToPlainObject, formatError } from '../utils';
+import { convertToPlainObject } from '../utils';
 import { LATEST_PRODUCTS_LIMIT, PAGE_SIZE } from '../constants';
-import { revalidatePath } from 'next/cache';
-import { insertProductSchema, updateProductSchema } from '../validators';
-import { z } from 'zod';
 import { Prisma } from '@prisma/client';
 
 // Get latest products
@@ -123,72 +120,6 @@ export async function getAllProducts({
   };
 }
 
-// Delete a product
-export async function deleteProduct(id: string) {
-  try {
-    const productExists = await prisma.product.findFirst({
-      where: { id },
-    });
-
-    if (!productExists) throw new Error('Product not found');
-
-    await prisma.product.delete({ where: { id } });
-
-    revalidatePath('/admin/products');
-
-    return {
-      success: true,
-      message: 'Product deleted successfully',
-    };
-  } catch (error) {
-    return { success: false, message: formatError(error) };
-  }
-}
-
-// Create a product
-export async function createProduct(data: z.infer<typeof insertProductSchema>) {
-  try {
-    const product = insertProductSchema.parse(data);
-    await prisma.product.create({ data: product });
-
-    revalidatePath('/admin/products');
-
-    return {
-      success: true,
-      message: 'Product created successfully',
-    };
-  } catch (error) {
-    return { success: false, message: formatError(error) };
-  }
-}
-
-// Update a product
-export async function updateProduct(data: z.infer<typeof updateProductSchema> & { id: string }) {
-  try {
-    const { id, ...updateData } = data;
-    const product = updateProductSchema.parse(updateData);
-    
-    const productExists = await prisma.product.findFirst({
-      where: { id },
-    });
-
-    if (!productExists) throw new Error('Product not found');
-
-    await prisma.product.update({
-      where: { id },
-      data: product,
-    });
-
-    revalidatePath('/admin/products');
-
-    return {
-      success: true,
-      message: 'Product updated successfully',
-    };
-  } catch (error) {
-    return { success: false, message: formatError(error) };
-  }
-}
 
 // Get all categories with product count
 export async function getAllCategories() {
